StationController & SensorType added
Set list sensors not functional

+++ ClimateChecker/src/main/java/Ontdekstation013/ClimateChecker/models/dto/sensorDto.java
@@ -7,9 +7,9 @@ import lombok.Setter;
 @Setter
 public class sensorDto extends Dto{
 
-    int id;
-    int stationId;
-    int typeId;
+    Long id;
+    Long stationId;
+    Long typeId;
     float data;
 
 }
